# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin 的快速启动项目",
        "title": "Gin Admin",
        "contact": {},
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/v1/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Query role list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Display name of role",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagination index. default(1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "disabled",
                            "enabled",
                            ""
                        ],
                        "type": "string",
                        "description": "Status of role (disabled, enabled",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "total number of items",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include menu IDs",
                        "name": "withMenus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResultList-po_Role"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Create role record",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RoleCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Get role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_Role"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Update role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RoleUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Delete role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.List-po_Role": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Role"
                    }
                },
                "limit": {
                    "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                    "type": "integer"
                },
                "page": {
                    "description": "pagination index. default(1)",
                    "type": "integer"
                },
                "total": {
                    "description": "total number of items",
                    "type": "integer"
                }
            }
        },
...
        "dto.Result-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.Result-po_Role": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/po.Role"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResultList-po_Role": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.List-po_Role"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.RoleCreateReq": {
            "type": "object",
            "required": [
                "code",
                "name",
                "status"
            ],
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "menuIds": {
                    "description": "Menu ids",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string",
                    "maxLength": 128
                },
                "rank": {
                    "description": "Rank for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (enabled, disabled)",
                    "type": "string",
                    "enum": [
                        "disabled",
                        "enabled"
                    ]
                }
            }
        },
        "dto.RoleUpdateReq": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "menuIds": {
                    "description": "Menu ids",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string",
                    "maxLength": 128
                },
                "rank": {
                    "description": "Rank for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (enabled, disabled)",
                    "type": "string",
                    "enum": [
                        "disabled",
                        "enabled"
                    ]
                }
            }
        },
...
        "po.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Child menus",
                    "allOf": [
                        {
                            "$ref": "#/definitions/po.Menus"
                        }
                    ]
                },
                "component": {
                    "description": "Component path of view",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "extra": {
                    "description": "Extra data for frontend",
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "method": {
                    "description": "Http method of resource",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string"
                },
                "parentId": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting (Order by desc)",
                    "type": "integer"
                },
                "redirect": {
                    "description": "Redirect path of menu",
                    "type": "string"
                },
                "roles": {
                    "$ref": "#/definitions/po.Roles"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string"
                },
                "title": {
                    "description": "Menu title",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (catalog, menu, button)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "po.Menus": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Menu"
            }
        },
        "po.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Menu"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (disabled, enabled)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.User"
                    }
                }
            }
        },
        "po.Roles": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Role"
            }
        },
        "po.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "Avatar URL",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about user",
                    "type": "string"
                },
                "email": {
                    "description": "Email of user",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string"
                },
                "realName": {
                    "description": "Real name of user",
                    "type": "string"
                },
                "roles": {
                    "description": "Roles of user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Role"
                    }
                },
                "status": {
                    "description": "Status of user (activated, freezed)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "username": {
                    "description": "Username for login",
                    "type": "string"
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}